requirements.txt
bs4
python-dotenv
requests
selenium